requirements.txt
blinker==1.7.0
boto3==1.34.47
botocore==1.34.47
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.0.7
Werkzeug==3.0.1
